Updated confusion matrix to not crash due to memory. Note bug with apparently water confused with seagrass spectra..

+++ scripts/training.py
@@ -27,36 +27,13 @@ def train_classifier(
     and map the UAV classes to the satellite classes. Train a random
     forest classifier and return."""
 
-    print(f"\tLoad in sites: {training_sites}")
-    training_dataframe = []
-    for training_site in training_sites:
-        training_file = samples_path / f"{training_site}_training_data.csv"
-        training_dataframe.append(pandas.read_csv(training_file))
-    training_dataframe = pandas.concat(training_dataframe, ignore_index=True)
-
-    print("\tMap UAV training ids to the specified satellite training ids")
-    uav_training_labels = (
-        pandas.read_csv(uav_labels_file, sep="\t", header=None, names=["Value", "Key"])
-        .set_index("Key")["Value"]
-        .to_dict()
-    )
-
-    # Drop UAV image specific classes - e.g. Shadow, Glare
-    class_ids_to_ignore = [uav_training_labels[key] for key in uav_classes_to_ignore]
-    training_dataframe = training_dataframe[
-        ~training_dataframe["uav_class_id"].isin(class_ids_to_ignore)
-    ]
-
-    # Map the UAV classes to those used in the satellite imagery
-    training_dataframe["satellite_class_id"] = training_dataframe["uav_class_id"]
-    for key in satellite_from_uav_classes.keys():
-        class_ids_to_map = [
-            uav_training_labels[key] for key in satellite_from_uav_classes[key]
-        ]
-        training_dataframe.loc[
-            training_dataframe["uav_class_id"].isin(class_ids_to_map),
-            "satellite_class_id",
-        ] = satellite_classes[key]
+    training_dataframe = map_satellite_ids_into_samples(
+        training_sites=training_sites,
+        samples_path=samples_path,
+        uav_labels_file=uav_labels_file,
+        uav_classes_to_ignore=uav_classes_to_ignore,
+        satellite_classes=satellite_classes,
+        satellite_from_uav_classes=satellite_from_uav_classes)
 
     print("\tTrain a Random Forest Model")
     training_classes = numpy.array(training_dataframe["satellite_class_id"])
@@ -67,6 +44,62 @@ def train_classifier(
     model = classifier.fit(training_observations, training_classes)
 
     return model, training_dataframe
+
+
+def load_samples(
+    training_sites: list,
+    samples_path: pathlib.Path,
+) -> pandas.DataFrame:
+    """Combine all training datasets and return."""
+
+    print(f"\tLoad in sites: {training_sites}")
+    samples_dataframe = []
+    for training_site in training_sites:
+        samples_file = samples_path / f"{training_site}_training_data.csv"
+        samples_dataframe.append(pandas.read_csv(samples_file))
+    samples_dataframe = pandas.concat(samples_dataframe, ignore_index=True)
+
+    return samples_dataframe
+
+
+def map_satellite_ids_into_samples(
+    training_sites: list,
+    samples_path: pathlib.Path,
+    uav_labels_file: pathlib.Path,
+    uav_classes_to_ignore: dict,
+    satellite_classes: dict,
+    satellite_from_uav_classes: dict,
+):
+    """Combine all training datasets then drop the UAV classes to ignore
+    and map the UAV classes to the satellite classes."""
+
+    samples_dataframe = load_samples(training_sites=training_sites,
+                                      samples_path=samples_path)
+
+    print("\tMap UAV training ids to the specified satellite training ids")
+    uav_training_labels = (
+        pandas.read_csv(uav_labels_file, sep="\t", header=None, names=["Value", "Key"])
+        .set_index("Key")["Value"]
+        .to_dict()
+    )
+
+    # Drop UAV image specific classes - e.g. Shadow, Glare
+    class_ids_to_ignore = [uav_training_labels[key] for key in uav_classes_to_ignore]
+    samples_dataframe = samples_dataframe[
+        ~samples_dataframe["uav_class_id"].isin(class_ids_to_ignore)
+    ]
+
+    # Map the UAV classes to those used in the satellite imagery
+    samples_dataframe["satellite_class_id"] = samples_dataframe["uav_class_id"]
+    for key in satellite_from_uav_classes.keys():
+        class_ids_to_map = [
+            uav_training_labels[key] for key in satellite_from_uav_classes[key]
+        ]
+        samples_dataframe.loc[
+            samples_dataframe["uav_class_id"].isin(class_ids_to_map),
+            "satellite_class_id",
+        ] = satellite_classes[key]
+    return samples_dataframe
 
 
 def predict_site(
@@ -145,10 +178,10 @@ def confusion_matrix_of_site(
         .to_dict()
     )
 
-    print("\tMatch satellite resolution to UAV then compare predictions to UAV")
+    print("Match satellite resolution to UAV then compare predictions to UAV")
 
     # Load in images
-    uav_training_data = utils.load_classification(filename=test_uav_file, chunks=True)
+    uav_training_data = utils.load_classification(filename=test_uav_file, chunks=True, masked=False)
     sat_prediction_data = utils.load_satellite(filename=prediction_file)
     uav_polygon = geopandas.read_file(polygon_file)
 
@@ -173,10 +206,7 @@ def confusion_matrix_of_site(
     del uav_training_data
     gc.collect()
 
-    # Ensure nan for no data and clip both to polygon
-    uav_training_data_reclassed.rio.set_nodata(numpy.nan)
-    sat_prediction_data.rio.set_nodata(numpy.nan)
-
+    # Clip both to polygon
     uav_training_data_reclassed = uav_training_data_reclassed.rio.clip(
         uav_polygon.geometry, all_touched=True, drop=True
     )
@@ -190,25 +220,20 @@ def confusion_matrix_of_site(
     )
 
     # Pull out the predictions vs ground truth
-    print("\tLoad UAV image")
     all_truth = []
     all_predictions = []
-    uav_training_data_reclassed.load()
     for time_index in range(len(sat_prediction_data["time"])):
 
         plot_filename = prediction_file.with_name(
             f"{prediction_file.stem}_confusion_matrix_time_{time_index}.png"
         )
 
-        if plot_filename.exists():
-            print(f"{plot_filename.name} exists. Skipping. Delete if you want regenerated.")
-            continue
-        
-        print(f"\tConstruct confusion matrix for time index: {time_index}")
-        truth = uav_training_data_reclassed.data[
-            sat_prediction_data.isel(time=0).notnull()
+        print(f"Time index: {time_index}")
+        print(f"\tExtract truth and predictions")
+        truth = uav_training_data_reclassed.values[
+            sat_prediction_data.isel(time=time_index).notnull()
         ]
-        predictions = sat_prediction_data.isel(time=time_index).data[
+        predictions = sat_prediction_data.isel(time=time_index).values[
             sat_prediction_data.isel(time=time_index).notnull()
         ]
         # drop any NaN interpolated from clipped areas in the prediction
@@ -219,25 +244,30 @@ def confusion_matrix_of_site(
         )
         truth = truth[mask]
         predictions = predictions[mask]
+        all_truth.append(truth)
+        all_predictions.append(predictions)
 
+        if plot_filename.exists():
+            print(f"{plot_filename.name} exists. Skipping. Delete if you want regenerated.")
+            continue
+        print(f"\tConstruct confusion matrix for time index")
         confusion_matrix = sklearn.metrics.confusion_matrix(
             truth, predictions, normalize="true"
         )
         figure = sklearn.metrics.ConfusionMatrixDisplay(
             confusion_matrix=confusion_matrix, display_labels=satellite_classes.keys()
         )
+
         figure.plot(cmap=matplotlib.pyplot.cm.Blues)
         matplotlib.pyplot.savefig(plot_filename, dpi=300, )
         matplotlib.pyplot.close()
-        all_truth.append(truth)
-        all_predictions.append(predictions)
 
     # Force free memory
     del uav_training_data_reclassed
     del sat_prediction_data
     gc.collect()
 
-    print("\tOverall confusion matrix across prediction dates")
+    print("Overall confusion matrix across prediction dates")
     all_truth = numpy.concatenate(all_truth)
     all_predictions = numpy.concatenate(all_predictions)
     confusion_matrix = sklearn.metrics.confusion_matrix(
@@ -249,7 +279,101 @@ def confusion_matrix_of_site(
     figure.plot(cmap=matplotlib.pyplot.cm.Blues)
     matplotlib.pyplot.savefig(overall_plot_filename, dpi=300, )
 
+    return all_truth, all_predictions
 
+
+def plot_model_feature_importance(training_dataframe, model_file):
+    """Plot the feature importance of the trained random forest model."""
+
+    plot_filename = model_file.with_name(f"{model_file.stem}_random_forest_feature_importance.png")
+    if plot_filename.exists():
+        print(f"{plot_filename.name} already exists. Delete if you've updated the "
+              "model and want to regenerate")
+    else:
+        model = joblib.load(model_file)
+        importance_df = pandas.DataFrame(
+            {'Feature': training_dataframe.drop(columns=["satellite_class_id", "uav_class_id", "time"]).columns,
+             'Importance': model.feature_importances_})
+        importance_df.sort_values(by='Importance', ascending=False).plot(kind='bar', x='Feature', y='Importance')
+        matplotlib.pyplot.savefig(model_file.with_name(f"{model_file.stem}_random_forest_feature_importance.png"), dpi=300)
+
+def plot_uav_classes(training_dataframe, uav_labels_file):
+    """Plot the UAV classes in the samples dataframe and return the figure"""
+    uav_training_labels = (
+        pandas.read_csv(uav_labels_file, sep="\t", header=None, names=["Value", "Key"])
+        .set_index("Key")["Value"]
+        .to_dict()
+    )
+    y_limits=(0, 6500)
+
+    # Plot satellite bands for UAV classes
+    number_uav_classes = len(training_dataframe["uav_class_id"].unique())
+    nrows = int(numpy.ceil(number_uav_classes/3))
+    figure, axes = matplotlib.pyplot.subplots(nrows=nrows, ncols=3, figsize=(21, 6*nrows))
+
+    for i, (class_id, ax) in enumerate(zip(training_dataframe["uav_class_id"].unique(), axes.flat)):
+
+        class_name = next((key for key, value in uav_training_labels.items() if value == class_id), None)
+
+        training_dataframe[training_dataframe["uav_class_id"] == class_id].drop(columns=["SCL", "uav_class_id", "satellite_class_id"]).plot(kind='box', ax=ax, ylim=y_limits)
+        ax.set_title(f"Spectral plot for class ID {class_name}")
+    return figure
+
+def plot_satellite_classes(training_dataframe, satellite_labels):
+    """Plot the satellite classes in the samples dataframe and return the figure"""
+
+    y_limits=(0, 6500)
+
+    # Plot satellite bands for satellite classes
+    nrows = int(numpy.ceil(len(satellite_labels)/3))
+    figure, axes = matplotlib.pyplot.subplots(nrows=nrows, ncols=3, figsize=(21, 6*nrows))
+    for i, (class_name, ax) in enumerate(zip(satellite_labels.keys(), axes.flat)):
+
+        class_id = satellite_labels[class_name]
+
+        training_dataframe[training_dataframe["satellite_class_id"] == class_id].drop(columns=["SCL", "satellite_class_id", "uav_class_id"]).plot(kind='box', ax=ax, ylim=y_limits)
+        ax.set_title(f"Spectral plot for class ID {class_name}")
+    return figure
+
+
+def save_samples_uav_classes(plot_filename, training_dataframe, uav_labels_file):
+    """Plot and save the UAV classes in the samples dataframe and return the figure"""
+    figure = plot_uav_classes(training_dataframe=training_dataframe, uav_labels_file=uav_labels_file)
+    figure.savefig(plot_filename, dpi=300)
+
+
+def save_samples_satellite_classes(plot_filename, training_dataframe, satellite_labels):
+    """Plot and save the satellite classes in the samples dataframe and return the figure"""
+    figure = plot_satellite_classes(training_dataframe=training_dataframe, satellite_labels=satellite_labels)
+    figure.savefig(plot_filename, dpi=300)
+
+
+def plot_training_data_class_distribution(training_dataframe, model_file, uav_labels_file, satellite_labels):
+    """Plot the class distribution of the training data."""
+
+    # Plot satellite bands for UAV classes
+    plot_filename = model_file.with_name(f"{model_file.stem}_training_uav_class_IDs.png")
+    if plot_filename.exists():
+        print(f"{plot_filename.name} already exists. Delete if you've updated the model"
+              " and want to regenerate")
+    else:
+        save_samples_uav_classes(plot_filename=plot_filename,
+                                 training_dataframe=training_dataframe,
+                                 uav_labels_file=uav_labels_file)
+
+    # Plot satellite bands for the satellite class used for prediction
+    plot_filename = model_file.with_name(f"{model_file.stem}_training_satellite_class_IDs.png")
+    if plot_filename.exists():
+        print(f"{plot_filename.name} already exists. Delete if you've updated the model"
+              " and want to regenerate")
+    else:
+        save_samples_satellite_classes(plot_filename=plot_filename,
+                                 training_dataframe=training_dataframe,
+                                 satellite_labels=satellite_labels)
+
+
+
+# Not implemented yet
 def confusion_matrix_of_site_satellite_resolution(
     test_uav_file,
     uav_labels_file,
@@ -304,10 +428,7 @@ def confusion_matrix_of_site_satellite_resolution(
     del uav_training_data
     gc.collect()
 
-    # Ensure nan for no data and clip both to polygon
-    uav_training_data_reclassed.rio.set_nodata(numpy.nan)
-    sat_prediction_data.rio.set_nodata(numpy.nan)
-
+    # Clip both to polygon
     uav_training_data_reclassed = uav_training_data_reclassed.rio.clip(
         uav_polygon.geometry, all_touched=True, drop=True
     )
@@ -341,15 +462,11 @@ def confusion_matrix_of_site_satellite_resolution(
             f"{prediction_file.stem}_confusion_matrix_{sentinel2.S2_RESOLUTION}_resolution_time_{time_index}.png"
         )
 
-        if plot_filename.exists():
-            print(f"{plot_filename.name} exists. Skipping. Delete if you want regenerated.")
-            continue
-
         print(f"\tConstruct confusion matrix for time index: {time_index}")
-        truth = uav_training_data_reclassed.data[
-            sat_prediction_data.isel(time=0).notnull()
+        truth = uav_training_data_reclassed.values[
+            sat_prediction_data.isel(time=time_index).notnull()
         ]
-        predictions = sat_prediction_data.isel(time=time_index).data[
+        predictions = sat_prediction_data.isel(time=time_index).values[
             sat_prediction_data.isel(time=time_index).notnull()
         ]
         # drop any NaN interpolated from clipped areas in the prediction
@@ -361,6 +478,12 @@ def confusion_matrix_of_site_satellite_resolution(
         truth = truth[mask]
         predictions = predictions[mask]
 
+        all_truth.append(truth)
+        all_predictions.append(predictions)
+
+        if plot_filename.exists():
+            print(f"{plot_filename.name} exists. Skipping. Delete if you want regenerated.")
+            continue
         confusion_matrix = sklearn.metrics.confusion_matrix(
             truth, predictions, normalize="true"
         )
@@ -370,8 +493,6 @@ def confusion_matrix_of_site_satellite_resolution(
         figure.plot(cmap=matplotlib.pyplot.cm.Blues)
         matplotlib.pyplot.savefig(plot_filename, dpi=300, )
         matplotlib.pyplot.close()
-        all_truth.append(truth)
-        all_predictions.append(predictions)
 
     # Force free memory
     del uav_training_data_reclassed
@@ -389,49 +510,3 @@ def confusion_matrix_of_site_satellite_resolution(
     )
     figure.plot(cmap=matplotlib.pyplot.cm.Blues)
     matplotlib.pyplot.savefig(overall_plot_filename, dpi=300, )
-
-
-def plot_model_feature_importance(training_dataframe, model_file):
-    """Plot the feature importance of the trained random forest model."""
-
-    plot_filename = model_file.with_name(f"{model_file.stem}_random_forest_feature_importance.png")
-    if plot_filename.exists():
-        print(f"{plot_filename.name} already exists. Delete if you've updated the "
-              "model and want to regenerate")
-    else:
-        model = joblib.load(model_file)
-        importance_df = pandas.DataFrame(
-            {'Feature': training_dataframe.drop(columns=["satellite_class_id", "uav_class_id", "time"]).columns,
-             'Importance': model.feature_importances_})
-        importance_df.sort_values(by='Importance', ascending=False).plot(kind='bar', x='Feature', y='Importance')
-        matplotlib.pyplot.savefig(model_file.with_name(f"{model_file.stem}_random_forest_feature_importance.png"), dpi=300)
-
-
-def plot_training_data_class_distribution(training_dataframe, model_file):
-    """Plot the class distribution of the training data."""
-    
-    # Plot satellite bands for UAV classes
-    plot_filename = model_file.with_name(f"{model_file.stem}_training_uav_class_IDs.png")
-    if plot_filename.exists():
-        print(f"{plot_filename.name} already exists. Delete if you've updated the model"
-              " and want to regenerate")
-    else:
-        number_uav_classes = len(training_dataframe["uav_class_id"].unique())
-        nrows = int(numpy.ceil(number_uav_classes/3))
-        figure, axes = matplotlib.pyplot.subplots(nrows=nrows, ncols=3, figsize=(21, 6*nrows))
-        for i, (class_id, ax) in enumerate(zip(training_dataframe["uav_class_id"].unique(), axes.flat)):
-            training_dataframe[training_dataframe["uav_class_id"] == 1].drop(columns=["SCL", "uav_class_id", "satellite_class_id"]).boxplot(ax=ax)  
-            ax.set_title(f"Spectral plot for class ID {class_id}")
-        matplotlib.pyplot.savefig(plot_filename, dpi=300)
-
-    # Plot satellite bands for the satellite class used for prediction
-    plot_filename = model_file.with_name(f"{model_file.stem}_training_satellite_class_IDs.png")
-    if plot_filename.exists():
-        print(f"{plot_filename.name} already exists. Delete if you've updated the model"
-              " and want to regenerate")
-    else:
-        figure, axes = matplotlib.pyplot.subplots(nrows=2, ncols=2, figsize=(14, 12))
-        for i, (class_id, ax) in enumerate(zip(training_dataframe["satellite_class_id"].unique(), axes.flat)):
-            training_dataframe[training_dataframe["satellite_class_id"] == 1].drop(columns=["SCL", "satellite_class_id", "uav_class_id"]).boxplot(ax=ax)  
-            ax.set_title(f"Spectral plot for class ID {class_id}")
-        matplotlib.pyplot.savefig(plot_filename, dpi=300)
